requirements.txt
requests==2.28.2
pytest==7.4.0
allure-pytest==2.15.0